+++ R/taxus_bin.R
@@ -1,15 +1,15 @@
 #' *Taxus baccata* forests
 #'
 #' A binary phytosociological table containing relevés of *Taxus baccata*
-#'   forests, in the northwest of the Iberian Peninsula.
+#'   forests, from the northwest of the Iberian Peninsula.
 #'
-#' @format A matrix with 209 rows and 33 columns. Each column correspond to a
-#'   phytosociological relevé and each row correspond to a taxon. Values in the
+#' @format A matrix with 209 rows and 33 columns. Each column corresponds to a
+#'   phytosociological relevé and each row corresponds to a taxon. Values in the
 #'   matrix denote presences (1) and absences (0).
 #'
 #' @source Portela-Pereira E., Monteiro-Henriques T., Casas C., Forner N.,
 #'   Garcia-Cabral I., Fonseca J.P. & Neto C. 2021. *Teixedos no noroeste da*
-#'   *Península Ibérica*. Finisterra 56(117): 127‐150.
+#'   *Península Ibérica*. Finisterra 56(117): 127-150.
 #'   \doi{10.18055/FINIS18102}.
 #'
 #' @examples