+++ NEWS.md
@@ -1,3 +1,17 @@
+# diffval 1.1.0
+
+## Major changes
+
+* Function bigdata_tdv() added, allowing Total Differential Value calculation
+  for big matrices, optionally using fork with package 'parallel'
+* Two auxiliary functions added to assist bigdata_tdv() in utils.R
+
+## Bug fixes
+
+* Improved sorting in tabulation() function
+* TDV calculation was corrected for some cases in optim_tdv_hill_climb() and in
+  the associated auxiliary function in utils.R
+
 # diffval 1.0.0
 
 * Updated version, for package release